Animation without layout
Tried over and over to animate within my layout, to no avail. Created a new template based on past class files and it seems I'm able to manipulate the coordinates of the container now, indicating there's something in my layout preventing that. By now, it's clearer than ever I won't make it in time for the due date, but I'll keep trying and see how far I can go with this....